requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
PyHamcrest==2.1.0
pytest==8.3.5
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
tomli==2.2.1
typing_extensions==4.13.0
urllib3==2.3.0
